# app/static/swagger.yaml
openapi: 3.0.1
info:
  title: Swagger Dr. Search
  description: This is an API document for Dr. Search App
  contact:
    email: info@smartlaboman.com
  version: 1.0.0
...
servers:
- url: https://syhaty-api.smartlaboman.com/project/api/v1
- url: http://127.0.0.1:5000/project/api/v1
security:
- basicAuth: []
- JWT: []
...
paths:
...
  /doctors/:
    get:
      tags:
      - doctor
      summary: Get All Doctors
      description: This can only be done by the logged in user.
      operationId: GetDoctors
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  doctors:
                    type: array
                    items:
                      $ref: '#/components/schemas/Doctor'
        401:
          $ref: '#/components/responses/Unauthorized'
  /doctors/{doctorId}:
    get:
      tags:
      - doctor
      summary: Get doctor
      description: This can only be done by the logged in user.
      operationId: getDoctorById
      parameters:
      - name: doctorId
        in: path
        description: The id that needs to get
        required: true
        schema:
          type: string
      responses:
        400:
          description: Invalid doctorId supplied
          content: {}
        404:
          description: Doctors not found
          content: {}
        401:
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags:
      - doctor
      summary: Updated Doctor
      description: This can only be done by the logged in user.
      operationId: updateDoctor
      parameters:
      - name: doctorId
        in: path
        description: ID that need to be updated
        required: true
        schema:
          type: string
      requestBody:
        description: Updated doctor object
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/Doctor'
        required: true
      responses:
        400:
          description: Invalid Doctor supplied
          content: {}
        404:
          description: Doctor not found
          content: {}
        401:
          $ref: '#/components/responses/Unauthorized'
      x-codegen-request-body-name: body
  /doctors/disable/{doctorId}:
    patch:
      tags:
      - doctor
      summary: Disable doctor
      description: This can only be done by the logged in user.
      operationId: disableDoctor
      parameters:
      - name: doctorId
        in: path
        description: The id that needs to be disable
        required: true
        schema:
          type: string
      responses:
        400:
          description: Invalid doctorId supplied
          content: {}
        404:
          description: User not found
          content: {}
        401:
          $ref: '#/components/responses/Unauthorized'
  /doctors/approve/{doctorId}:
    patch:
      tags:
      - doctor
      summary: Approve doctor
      description: This can only be done by the logged in user.
      operationId: approveDoctor
      parameters:
      - name: doctorId
        in: path
        description: The id that needs to be approved
        required: true
        schema:
          type: string
      responses:
        400:
          description: Invalid doctorId supplied
          content: {}
        404:
          description: User not found
          content: {}
        401:
          $ref: '#/components/responses/Unauthorized'
  /patients/:
    get:
      tags:
      - patient
      summary: Get All Patients
      description: This can only be done by the logged in user.
      operationId: GetPatients
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  patients:
                    type: array
                    items:
                      $ref: '#/components/schemas/Patient'
        401:
          $ref: '#/components/responses/Unauthorized'
  /patients/{patientId}:
    get:
      tags:
      - patient
      summary: Get patient
      description: This can only be done by the logged in user.
      operationId: getPatientById
      parameters:
      - name: patientId
        in: path
        description: The id that needs to get
        required: true
        schema:
          type: string
      responses:
        400:
          description: Invalid patientId supplied
          content: {}
        404:
          description: Patients not found
          content: {}
        401:
          $ref: '#/components/responses/Unauthorized'
...
  /centers/{centerId}:
    get:
      tags:
      - center
      summary: Get center
      description: This can only be done by the logged in user.
      operationId: getCenterById
      parameters:
      - name: centerId
        in: path
        description: The id that needs to get
        required: true
        schema:
          type: string
      responses:
        400:
          description: Invalid centerId supplied
          content: {}
        404:
          description: Centers not found
          content: {}
        401:
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags:
      - center
      summary: Updated Center
      description: This can only be done by the logged in user.
      operationId: updateCenter
      parameters:
      - name: centerId
        in: path
        description: ID that need to be updated
        required: true
        schema:
          type: string
      requestBody:
        description: Updated center object
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/Center'
        required: true
      responses:
        400:
          description: Invalid Center supplied
          content: {}
        404:
          description: Center not found
          content: {}
        401:
          $ref: '#/components/responses/Unauthorized'
      x-codegen-request-body-name: body
...
  /settings/:
    get:
      tags:
      - setting
      summary: Get All Settings
      description: This can only be done by the logged in user.
      operationId: GetSettings
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  settings:
                    type: array
                    items:
                      $ref: '#/components/schemas/Setting'
        401:
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
      - setting
      summary: Add Setting
      description: This can only be done by the logged in user.
      operationId: AddSetting
      requestBody:
        description: Add setting object
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/Setting'
        required: true
      responses:
        400:
          description: Invalid setting supplied
          content: {}
        401:
          $ref: '#/components/responses/Unauthorized'
      x-codegen-request-body-name: body
  /settings/{key}:
    get:
      tags:
      - setting
      summary: Get setting
      description: This can only be done by the logged in user.
      operationId: getSettingBykey
      parameters:
      - name: key
        in: path
        description: The id that needs to get
        required: true
        schema:
          type: string
      responses:
        400:
          description: Invalid key supplied
          content: {}
        404:
          description: Settings not found
          content: {}
        401:
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags:
      - setting
      summary: Delete setting
      description: This can only be done by the logged in user.
      operationId: deleteSetting
      parameters:
      - name: key
        in: path
        description: The id that needs to be deleted
        required: true
        schema:
          type: string
      responses:
        400:
          description: Invalid key supplied
          content: {}
        404:
          description: Setting not found
          content: {}
        401:
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags:
      - setting
      summary: Updated Setting
      description: This can only be done by the logged in user.
      operationId: updateSetting
      parameters:
      - name: key
        in: path
        description: ID that need to be updated
        required: true
        schema:
          type: string
      requestBody:
        description: Updated setting object
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/Setting'
        required: true
      responses:
        400:
          description: Invalid setting supplied
          content: {}
        404:
          description: setting not found
          content: {}
        401:
          $ref: '#/components/responses/Unauthorized'
      x-codegen-request-body-name: body
  /settings/category/{category}:
    get:
      tags:
      - setting
      summary: Get setting
      description: This can only be done by the logged in user.
      operationId: getSettingByCategory
      parameters:
      - name: category
        in: path
        description: The model type that needs to get
        required: true
        schema:
          type: string
      responses:
        400:
          description: Invalid category supplied
          content: {}
        404:
          description: Settings not found
          content: {}
        401:
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Center:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        country:
          type: string
          default: OM
        address:
          type: string
        place_location:
          type: string
        exact_location:
          type: string
    Patient:
      type: object
      properties:
        id:
          type: integer
          format: int64
        first_name:
          type: string
          example: Ali
        last_name:
          type: string
          example: Al
        gender:
          type: string
        birth_date:
          type: string
          format: date-time
        blood_type:
          type: string
        country:
          type: string
          default: OM
    Doctor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        public_id:
          type: string
          format: uuid
        first_name:
          type: string
          example: Ali
        last_name:
          type: string
          example: Al
        gender:
          type: string
        birth_date:
          type: string
          format: date-time
        blood_type:
          type: string
        country:
          type: string
          default: OM
        address:
          type: string
        place_location:
          type: string
        exact_location:
          type: string
        speciality:
          type: string
          format: JSON
        professional_experiance:
          type: string
        certification:
          type: string
          format: JSON
        experiance:
          type: string
        education:
          type: string
          format: JSON
        memberships:
          type: string
          format: JSON
        follow_up_fee:
          type: number
          format: float
        consultation_fee:
          type: number
          format: float
        biography:
          type: string
        medical_board_registrations:
          type: string
          format: JSON
        know_languages:
          type: string
          format: JSON
...
    Setting:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
        category:
          type: string
        last_modified:
          type: string
          format: date-time
...
    Error:
      required:
      #- code
      - message
      type: object
      properties:
        #code:
        #  type: string
        message:
          type: string
          example: Authorization information is missing or invalid.
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    JWT:
      type: apiKey
      name: x-access-tokens
      in: header